organized code fixed bug that was not capturing elements that had a base path of value[x] moved fixupContentReferenceSpecifier functionality to the FHIRClassInfoBuilder
