+++ src/main/java/org/opencds/cqf/modelinfo/fhir/FHIRClassInfoBuilder.java
@@ -4,7 +4,11 @@ import java.util.Collection;
 import java.util.Map;
 
 import org.hl7.elm_modelinfo.r1.ClassInfo;
+import org.hl7.elm_modelinfo.r1.ClassInfoElement;
+import org.hl7.elm_modelinfo.r1.ListTypeSpecifier;
+import org.hl7.elm_modelinfo.r1.NamedTypeSpecifier;
 import org.hl7.elm_modelinfo.r1.TypeInfo;
+import org.hl7.elm_modelinfo.r1.TypeSpecifier;
 import org.hl7.fhir.r4.model.StructureDefinition;
 import org.hl7.fhir.r4.model.StructureDefinition.StructureDefinitionKind;
 import org.hl7.fhir.r4.model.StructureDefinition.TypeDerivationRule;
@@ -39,5 +43,83 @@ public class FHIRClassInfoBuilder extends ClassInfoBuilder {
             .filter(y -> this.hasContentReferenceTypeSpecifier(y))
             .forEach(y -> this.fixupContentReferenceSpecifier("FHIR", y))
         );
+    }
+
+    private TypeSpecifier resolveContentReference(String modelName, String path) throws Exception {
+        String root = this.getQualifier(this.stripRoot(path, "#"));
+        String elementPath = this.unQualify(this.stripRoot(path, "#"));
+
+        TypeInfo rootType = this.resolveType(modelName + "." + root);
+        ClassInfoElement element = this.forPath(((ClassInfo) rootType).getElement(), elementPath);
+        return this.getTypeSpecifier(element);
+    }
+
+    private Boolean isContentReferenceTypeSpecifier(TypeSpecifier typeSpecifier) {
+        if (typeSpecifier instanceof NamedTypeSpecifier) {
+            NamedTypeSpecifier nts = (NamedTypeSpecifier) typeSpecifier;
+            return nts.getName().startsWith("#");
+        } else if (typeSpecifier instanceof ListTypeSpecifier) {
+            ListTypeSpecifier lts = (ListTypeSpecifier) typeSpecifier;
+            if (lts.getElementType().startsWith("#")) {
+                return true;
+            } else if (lts.getElementTypeSpecifier() != null) {
+                return this.isContentReferenceTypeSpecifier(lts.getElementTypeSpecifier());
+            }
+        }
+
+        return false;
+    }
+
+    private String getContentReference(TypeSpecifier typeSpecifier) {
+        if (typeSpecifier instanceof NamedTypeSpecifier) {
+            NamedTypeSpecifier nts = (NamedTypeSpecifier) typeSpecifier;
+            if (nts.getName().startsWith("#")) {
+                return nts.getName();
+            }
+        } else if (typeSpecifier instanceof ListTypeSpecifier) {
+            ListTypeSpecifier lts = (ListTypeSpecifier) typeSpecifier;
+            if (lts.getElementType().startsWith("#")) {
+                return lts.getElementType();
+            } else if (lts.getElementTypeSpecifier() != null) {
+                return this.getContentReference(lts.getElementTypeSpecifier());
+            }
+        }
+
+        return null;
+    }
+
+    protected Boolean hasContentReferenceTypeSpecifier(ClassInfoElement element) {
+
+        return element.getElementType() != null && element.getElementType().startsWith("#")
+                || this.isContentReferenceTypeSpecifier(element.getElementTypeSpecifier());
+    }
+
+    protected ClassInfoElement fixupContentReferenceSpecifier(String modelName, ClassInfoElement element) {
+        ClassInfoElement result = null;
+        try {
+            if (this.hasContentReferenceTypeSpecifier(element)) {
+                result = element;
+                if (element.getElementType() != null && element.getElementType().startsWith("#")) {
+                    element.setElementType(this.getTypeName(
+                            (NamedTypeSpecifier) this.resolveContentReference(modelName, element.getElementType())));
+                } else if (element.getElementTypeSpecifier() != null
+                        && element.getElementTypeSpecifier() instanceof ListTypeSpecifier) {
+                    ListTypeSpecifier lts = new ListTypeSpecifier();
+                    lts.setElementTypeSpecifier(this.resolveContentReference(modelName,
+                            this.getContentReference(element.getElementTypeSpecifier())));
+
+                    element.setElementTypeSpecifier(lts);
+                } else if (element.getElementTypeSpecifier() != null) {
+                    element.setElementTypeSpecifier(this.resolveContentReference(modelName,
+                            this.getContentReference(element.getElementTypeSpecifier())));
+                } else
+                    return element;
+            }
+        } catch (Exception e) {
+            System.out.println("Error fixing up contentreferencetypespecifier: " + e.getMessage());
+            e.printStackTrace();
+        }
+
+        return result;
     }
 }
